+++ src/web/assets/JsApiAsset.php
@@ -12,6 +12,7 @@
 namespace doublesecretagency\googlemaps\web\assets;
 
 use craft\web\AssetBundle;
+use doublesecretagency\googlemaps\GoogleMapsPlugin;
 
 /**
  * Class JsApiAsset
@@ -32,9 +33,16 @@ class JsApiAsset extends AssetBundle
             GoogleMapsAsset::class,
         ];
 
+        // Whether to use minified JavaScript files
+        $minifyJsFiles = (GoogleMapsPlugin::$plugin->getSettings()->minifyJsFiles ?? false);
+
+        // Optionally use minified files
+        $min = ($minifyJsFiles ? 'min.' : '');
+
+        // Load JS files
         $this->js = [
-            'js/googlemaps.js',
-            'js/dynamicmap.js',
+            "js/googlemaps.{$min}js",
+            "js/dynamicmap.{$min}js",
         ];
     }